refactor(natural-language-classifier): remove compatibility layer for natural language classifier
BREAKING CHANGE: Deprecated methods in Natural Language Classifier are no longer available. Changed parameter names are no longer interally corrected.

To migrate your code, use the methods and parameters currently available with the service as documented here: https://www.ibm.com/watson/developercloud/natural-language-classifier/api/v1/node.html?node
